add sample and base strategy classes

+++ src/event/FillEvent.hpp
@@ -6,7 +6,7 @@
 struct FillEvent : Event {
 
   FillEvent(int timeindex, std::string symbol, std::string exchange,
-            int quantity, std::string direction, int fillCost, int commission)
+            float quantity, std::string direction, int fillCost, int commission)
       : Event("Fill"), timeindex(timeindex), symbol(symbol), exchange(exchange),
         quantity(quantity), direction(direction), fillCost(fillCost),
         commission(commission){};
@@ -16,7 +16,7 @@ struct FillEvent : Event {
   int timeindex;         // The bar-resolution when the order was filled
   std::string symbol;    // The instrument which was filled
   std::string exchange;  // The exchange where the order was filled
-  int quantity;          // The filled quantity
+  float quantity;        // The filled quantity
   std::string direction; // The direction of fill (’BUY’ or ’SELL’)
   int fillCost;          // The holdings value in dollars.
                          //! (not known during backtesting)
