+++ src/event/OrderEvent.hpp
@@ -5,7 +5,7 @@
 
 struct OrderEvent : Event {
 
-  OrderEvent(std::string symbol, std::string orderType, int quantity,
+  OrderEvent(std::string symbol, std::string orderType, float quantity,
              std::string direction)
       : Event("Order"), symbol(symbol), orderType(orderType),
         quantity(quantity), direction(direction){};
@@ -13,7 +13,7 @@ struct OrderEvent : Event {
 
   std::string symbol;    // The instrument to trade.
   std::string orderType; // "MKT" or "LMT" for Market or Limit.
-  int quantity;          // Non-negative integer for quantity.
+  float quantity;        // Non-negative integer for quantity.
   std::string direction; // "BUY" or "SELL" for long or short.
 };
